+++ signify/authenticode/signed_file.py
@@ -5,9 +5,15 @@ import os
 from collections.abc import Iterable, Iterator
 from typing import TYPE_CHECKING, Any, BinaryIO, Literal
 
+from typing_extensions import Self
+
 from signify._typing import HashFunction
 from signify.asn1.hashing import ACCEPTED_DIGEST_ALGORITHMS
-from signify.exceptions import AuthenticodeNotSignedError, ParseError
+from signify.exceptions import (
+    AuthenticodeFingerprintNotProvidedError,
+    AuthenticodeNotSignedError,
+    ParseError,
+)
 from signify.x509 import Certificate
 
 if TYPE_CHECKING:
@@ -15,15 +21,9 @@ if TYPE_CHECKING:
 
 
 class AuthenticodeFile:
-    """File signed with Authenticode."""
-
-    def __init__(self, file_obj: BinaryIO):
-        """An Authenticode-signed file that is to be parsed to find the relevant
-        sections for Authenticode parsing.
-
-        :param file_obj: An Authenticode-signed file opened in binary file
-        """
-        raise NotImplementedError()
+    """An Authenticode-signed file that is to be parsed to find the relevant
+    sections for Authenticode parsing.
+    """
 
     @classmethod
     def detect(cls, file_obj: BinaryIO) -> AuthenticodeFile:
@@ -40,6 +40,9 @@ class AuthenticodeFile:
             from .signed_pe import SignedPEFile
 
             return SignedPEFile(file_obj)
+        elif header.startswith(b"PKCX"):
+            file_obj.seek(-4, os.SEEK_CUR)
+            return AuthenticodeSignedDataFile.from_envelope(file_obj.read())
 
         raise ParseError("Unknown file type.")
 
@@ -232,7 +235,10 @@ class AuthenticodeFile:
 
     def get_fingerprint(self, digest_algorithm: HashFunction) -> bytes:
         """Gets the fingerprint for this file"""
-        raise NotImplementedError
+        raise AuthenticodeFingerprintNotProvidedError(
+            f"Fingerprint for digest algorithm {digest_algorithm.__name__} could not "
+            "be calculated and was not provided as pre-calculated expected hash."
+        )
 
     def _calculate_expected_hashes(
         self,
@@ -256,3 +262,50 @@ class AuthenticodeFile:
             fingerprint = self.get_fingerprint(digest_algorithm)
             expected_hashes[digest_algorithm().name] = fingerprint
         return expected_hashes
+
+
+class AuthenticodeSignedDataFile(AuthenticodeFile):
+    """Simple transparent :class:`AuthenticodeFile` class that operates on an
+    already-parsed :class:`structures.AuthenticodeSignedData`. This can be used in
+    places where the parsed SignedData object is present, but the original file is no
+    longer present, or for parsing P7X files.
+
+    Note that the :meth:`get_fingerprint` method is not implemented, so all hashes
+    must be provided as expected hashes in the :meth:`verify` method.
+
+    Remember that you can directly manipulate :class:`structures.AuthenticodeSignedData`
+    objects and that this class is a very simple shim. If you don't need the features
+    provided by :class:`AuthenticodeFile`, simply use the
+    :class:`structures.AuthenticodeSignedData` directly.
+    """
+
+    def __init__(self, signed_data: structures.AuthenticodeSignedData | None) -> None:
+        """
+        :param signed_data: The signed data object we're operating on. Can be None
+            to allow filling it in later (see :meth:`from_envelope`).
+        """
+        self.signed_data = signed_data
+
+    @classmethod
+    def from_envelope(cls, data: bytes) -> Self:
+        """Creates a :class:`AuthenticodeSignedDataFile` from a data envelope. This
+        will instantiate an 'empty' :class:`AuthenticodeSignedDataFile` object, and
+        fill it in.
+        """
+        from signify.authenticode import structures
+
+        signed_file = cls(None)
+        signed_file.signed_data = structures.AuthenticodeSignedData.from_envelope(
+            data, signed_file=signed_file
+        )
+        return signed_file
+
+    def iter_signed_datas(
+        self, *, include_nested: bool = True, ignore_parse_errors: bool = True
+    ) -> Iterator[structures.AuthenticodeSignedData]:
+        if self.signed_data is None:
+            raise Exception("Object not instantiated yet.")
+        if include_nested:
+            yield from self.signed_data.iter_recursive_nested()
+        else:
+            yield self.signed_data