+++ signify/authenticode/signed_msi.py
@@ -89,14 +89,6 @@ class SignedMsiFile(AuthenticodeFile):
     def iter_signed_datas(
         self, *, include_nested: bool = True, ignore_parse_errors: bool = True
     ) -> Iterator[structures.AuthenticodeSignedData]:
-        def recursive_nested(
-            signed_data: structures.AuthenticodeSignedData,
-        ) -> Iterator[structures.AuthenticodeSignedData]:
-            yield signed_data
-            if include_nested:
-                for nested in signed_data.signer_info.nested_signed_datas:
-                    yield from recursive_nested(nested)
-
         try:
             if not self._ole_file.exists(DIGITAL_SIGNATURE_ENTRY_NAME):
                 raise AuthenticodeNotSignedError(
@@ -106,12 +98,13 @@ class SignedMsiFile(AuthenticodeFile):
             with self._ole_file.openstream(DIGITAL_SIGNATURE_ENTRY_NAME) as fh:
                 b_data = fh.read()
 
-            yield from recursive_nested(
-                structures.AuthenticodeSignedData.from_envelope(
-                    b_data, signed_file=self
-                )
+            signed_data = structures.AuthenticodeSignedData.from_envelope(
+                b_data, signed_file=self
             )
-
+            if include_nested:
+                yield from signed_data.iter_recursive_nested()
+            else:
+                yield signed_data
         except SignedMsiParseError:
             if not ignore_parse_errors:
                 raise